+++ src/main/java/com/cch/codechallengehub/web/exception/GlobalExceptionHandler.java
@@ -1,6 +1,4 @@
-package com.cch.codechallengehub.api.exception;
-
-import static com.cch.codechallengehub.api.exception.ErrorCode.UNKNOWN_ERROR;
+package com.cch.codechallengehub.web.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.annotation.Order;
@@ -11,13 +9,13 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 @Slf4j
 @RestControllerAdvice
 @Order(2)
-public class GlobalExceptionHandler extends MvcExceptionHandler {
+public class GlobalExceptionHandler {
 	@ExceptionHandler(RuntimeException.class)
 	public ResponseEntity<ErrorResponse<ErrorCode>> handleRuntimeException(RuntimeException ex) {
 
 		String message = ex.getMessage();
 		ErrorResponse<ErrorCode> response = ErrorResponse.<ErrorCode>builder()
-			.errorCode(UNKNOWN_ERROR)
+			.errorCode(ErrorCode.UNKNOWN_ERROR)
 			.msg(message)
 			.build();
 
@@ -29,7 +27,7 @@ public class GlobalExceptionHandler extends MvcExceptionHandler {
 
 		String message = ex.getMessage();
 		ErrorResponse<ErrorCode> response = ErrorResponse.<ErrorCode>builder()
-			.errorCode(UNKNOWN_ERROR)
+			.errorCode(ErrorCode.UNKNOWN_ERROR)
 			.msg(message)
 			.build();
