+++ src/test/java/com/cch/codechallengehub/web/exception/TestDto.java
@@ -1,4 +1,4 @@
-package com.cch.codechallengehub.api.exception;
+package com.cch.codechallengehub.web.exception;
 
 
 import jakarta.validation.constraints.NotNull;
